split mov_cond from misc
