+++ crates/core/executor/src/artifacts/mips_costs.json
@@ -41,10 +41,11 @@
   "SyscallCore": 22,
   "SysLinux": 125,
   "Bn254Fp2AddSubAssign": 1382,
+  "MovCond": 37,
   "Bls12381FpOpAssign": 1048,
   "Cpu": 119,
   "ShaCompress": 40480,
   "MemoryInstrs": 110,
-  "MiscInstrs": 151,
+  "MiscInstrs": 149,
   "Secp256k1DoubleAssign": 4492
 }
