+++ crates/core/executor/src/events/instr.rs
@@ -247,8 +247,6 @@ pub struct MiscEvent {
     pub next_pc: u32,
     /// The opcode.
     pub opcode: Opcode,
-    /// The register id for first operand value.
-    pub op_a: u8,
     /// The first operand value.
     pub a: u32,
     /// The second operand value.
@@ -271,13 +269,51 @@ impl MiscEvent {
         pc: u32,
         next_pc: u32,
         opcode: Opcode,
-        op_a: u8,
         a: u32,
         b: u32,
         c: u32,
         prev_a: u32,
         hi_record: MemoryWriteRecord,
     ) -> Self {
-        Self { clk, shard, pc, next_pc, opcode, op_a, a, b, c, prev_a, hi_record }
+        Self { clk, shard, pc, next_pc, opcode, a, b, c, prev_a, hi_record }
+    }
+}
+
+
+/// Misc Instruction Event.
+///
+/// This object encapsulated the information needed to prove a MIPS misc operation.
+#[derive(Debug, Clone, Copy, Serialize, Deserialize)]
+#[repr(C)]
+pub struct MovCondEvent {
+    /// The program counter.
+    pub pc: u32,
+    pub next_pc: u32,
+    /// The opcode.
+    pub opcode: Opcode,
+    /// The first operand value.
+    pub a: u32,
+    /// The second operand value.
+    pub b: u32,
+    /// The third operand value.
+    pub c: u32,
+    /// The third operand value.
+    pub prev_a: u32,
+}
+
+impl MovCondEvent {
+    /// Create a new [`MovCondEvent`].
+    #[must_use]
+    #[allow(clippy::too_many_arguments)]
+    pub fn new(
+        pc: u32,
+        next_pc: u32,
+        opcode: Opcode,
+        a: u32,
+        b: u32,
+        c: u32,
+        prev_a: u32,
+    ) -> Self {
+        Self { pc, next_pc, opcode, a, b, c, prev_a }
     }
 }
