+++ crates/stark/src/air/builder.rs
@@ -203,7 +203,7 @@ pub trait InstructionAirBuilder: BaseAirBuilder {
         hi: Word<impl Into<Self::Expr>>,
         op_a_immutable: impl Into<Self::Expr>,
         is_rw_a: impl Into<Self::Expr>,
-        is_write_hi: impl Into<Self::Expr>,
+        is_check_memory: impl Into<Self::Expr>,
         is_halt: impl Into<Self::Expr>,
         is_sequential: impl Into<Self::Expr>,
         multiplicity: impl Into<Self::Expr>,
@@ -221,7 +221,7 @@ pub trait InstructionAirBuilder: BaseAirBuilder {
             .chain(hi.0.into_iter().map(Into::into))
             .chain(once(op_a_immutable.into()))
             .chain(once(is_rw_a.into()))
-            .chain(once(is_write_hi.into()))
+            .chain(once(is_check_memory.into()))
             .chain(once(is_halt.into()))
             .chain(once(is_sequential.into()))
             .collect();
@@ -249,7 +249,7 @@ pub trait InstructionAirBuilder: BaseAirBuilder {
         hi: Word<impl Into<Self::Expr>>,
         op_a_immutable: impl Into<Self::Expr>,
         is_rw_a: impl Into<Self::Expr>,
-        is_write_hi: impl Into<Self::Expr>,
+        is_check_memory: impl Into<Self::Expr>,
         is_halt: impl Into<Self::Expr>,
         is_sequential: impl Into<Self::Expr>,
         multiplicity: impl Into<Self::Expr>,
@@ -267,7 +267,7 @@ pub trait InstructionAirBuilder: BaseAirBuilder {
             .chain(hi.0.into_iter().map(Into::into))
             .chain(once(op_a_immutable.into()))
             .chain(once(is_rw_a.into()))
-            .chain(once(is_write_hi.into()))
+            .chain(once(is_check_memory.into()))
             .chain(once(is_halt.into()))
             .chain(once(is_sequential.into()))
             .collect();
